Log structured object in database.
